Moved EndScoreboard to FinalActivity and updated tests
