+++ .idea/androidTestResultsUserPreferences.xml
@@ -341,7 +341,9 @@
             <AndroidTestResultsTableState>
               <option name="preferredColumnWidths">
                 <map>
+                  <entry key="CUBOT&#10; KINGKONG MINI2" value="120" />
                   <entry key="Duration" value="90" />
+                  <entry key="KKMN2211115031121" value="120" />
                   <entry key="Pixel_2_API_33" value="120" />
                   <entry key="Tests" value="360" />
                 </map>
@@ -415,19 +417,6 @@
             </AndroidTestResultsTableState>
           </value>
         </entry>
-        <entry key="1242169909">
-          <value>
-            <AndroidTestResultsTableState>
-              <option name="preferredColumnWidths">
-                <map>
-                  <entry key="Duration" value="90" />
-                  <entry key="Pixel_2_API_33" value="120" />
-                  <entry key="Tests" value="360" />
-                </map>
-              </option>
-            </AndroidTestResultsTableState>
-          </value>
-        </entry>
         <entry key="1121581564">
           <value>
             <AndroidTestResultsTableState>
@@ -436,6 +425,19 @@
                   <entry key="39V4C19703010730" value="120" />
                   <entry key="Duration" value="90" />
                   <entry key="HUAWEI&#10; POT-LX1" value="120" />
+                  <entry key="Tests" value="360" />
+                </map>
+              </option>
+            </AndroidTestResultsTableState>
+          </value>
+        </entry>
+        <entry key="1242169909">
+          <value>
+            <AndroidTestResultsTableState>
+              <option name="preferredColumnWidths">
+                <map>
+                  <entry key="Duration" value="90" />
+                  <entry key="Pixel_2_API_33" value="120" />
                   <entry key="Tests" value="360" />
                 </map>
               </option>
